+++ cli_tool_audit/freeze.py
@@ -6,50 +6,54 @@ import tempfile
 
 import cli_tool_audit.call_tools as call_tools
 from cli_tool_audit.config_manager import ConfigManager
+from cli_tool_audit.models import SchemaType
 
 
-def freeze_requirements(tool_names: list[str]) -> dict[str, call_tools.ToolAvailabilityResult]:
+def freeze_requirements(tool_names: list[str], schema: SchemaType) -> dict[str, call_tools.ToolAvailabilityResult]:
     """
     Capture the current version of a list of tools.
 
     Args:
         tool_names (list[str]): A list of tool names.
+        schema (SchemaType): The schema to use for the version.
 
     Returns:
         dict[str, call_tools.ToolAvailabilityResult]: A dictionary of tool names and versions.
     """
     results = {}
     for tool_name in tool_names:
-        result = call_tools.check_tool_availability(tool_name, "--version")
+        result = call_tools.check_tool_availability(tool_name, schema, "--version")
         results[tool_name] = result
     return results
 
 
-def freeze_to_config(tool_names: list[str], config_path: str) -> None:
+def freeze_to_config(tool_names: list[str], config_path: str, schema: SchemaType) -> None:
     """
     Capture the current version of a list of tools and write them to a config file.
 
     Args:
         tool_names (list[str]): A list of tool names.
         config_path (str): The path to the config file.
+        schema (SchemaType): The schema to use for the version.
     """
-    results = freeze_requirements(tool_names)
+    results = freeze_requirements(tool_names, schema=schema)
     config_manager = ConfigManager(config_path)
     config_manager.read_config()
     for tool_name, result in results.items():
         if result.is_available and result.version:
-            config_manager.create_update_tool_config(tool_name, {"version_snapshot": result.version})
+            config_manager.create_update_tool_config(tool_name, {"version": result.version})
 
 
-def freeze_to_screen(tool_names: list[str]) -> None:
+def freeze_to_screen(tool_names: list[str], schema: SchemaType) -> None:
     """
     Capture the current version of a list of tools, write them to a temp config file,
     and print the 'cli-tools' section of the config.
 
     Args:
         tool_names (list[str]): A list of tool names.
+        schema (SchemaType): The schema to use for the version.
     """
-    results = freeze_requirements(tool_names)
+    results = freeze_requirements(tool_names, schema=schema)
 
     # Create a temporary directory and file
     with tempfile.TemporaryDirectory() as temp_dir:
@@ -59,7 +63,7 @@ def freeze_to_screen(tool_names: list[str]) -> None:
 
         for tool_name, result in results.items():
             if result.is_available and result.version:
-                config_manager.create_update_tool_config(tool_name, {"version_snapshot": result.version})
+                config_manager.create_update_tool_config(tool_name, {"version": result.version})
 
         # Save the config
         # pylint: disable=protected-access
@@ -72,4 +76,4 @@ def freeze_to_screen(tool_names: list[str]) -> None:
 
 
 if __name__ == "__main__":
-    freeze_to_screen(["python", "pip", "poetry"])
+    freeze_to_screen(["python", "pip", "poetry"], schema=SchemaType.SNAPSHOT)