+++ cli_tool_audit/view_pipx_stress_test.py
@@ -9,6 +9,8 @@ from concurrent.futures import ThreadPoolExecutor
 from threading import Lock
 from typing import Any
 
+from tqdm import tqdm
+
 import cli_tool_audit.views as views
 from cli_tool_audit.models import CliToolConfig
 
@@ -82,16 +84,19 @@ def report_for_pipx_tools(max_count: int = -1) -> None:
     lock = Lock()
     with ThreadPoolExecutor(max_workers=num_cpus) as executor:
         # Submit tasks to the executor
-        futures = [
-            executor.submit(views.check_tool_wrapper, (tool, config, lock, enable_cache))
-            for tool, config in cli_tools.items()
-        ]
+        disable = len(cli_tools) < 5 or os.environ.get("CI") or os.environ.get("NO_COLOR")
+        with tqdm(total=len(cli_tools), disable=disable) as progress_bar:
+            futures = [
+                executor.submit(views.check_tool_wrapper, (tool, config, lock, enable_cache))
+                for tool, config in cli_tools.items()
+            ]
 
-        results = []
-        # Process the results as they are completed
-        for future in concurrent.futures.as_completed(futures):
-            result = future.result()
-            results.append(result)
+            results = []
+            # Process the results as they are completed
+            for future in concurrent.futures.as_completed(futures):
+                result = future.result()
+                tqdm.update(progress_bar, 1)
+                results.append(result)
         views.pretty_print_results(results)