+++ tests/test_hypothesis/test_version_parsing.py
@@ -1,10 +1,10 @@
 # This test code was written by the `hypothesis.extra.ghostwriter` module
 # and is provided under the Creative Commons Zero public domain dedication.
 
+from hypothesis import given
+from hypothesis import strategies as st
+
 import cli_tool_audit.version_parsing
-import packaging.version
-from hypothesis import given, strategies as st
-from packaging.version import Version
 
 # Random strings are not valid versions, duh.
 # @given(version=st.builds(Version, version=st.text()))
@@ -14,16 +14,12 @@ from packaging.version import Version
 
 @given(input_string=st.text())
 def test_fuzz_extract_first_semver_version(input_string: str) -> None:
-    cli_tool_audit.version_parsing.extract_first_semver_version(
-        input_string=input_string
-    )
+    cli_tool_audit.version_parsing.extract_first_semver_version(input_string=input_string)
 
 
 @given(input_string=st.text())
 def test_fuzz_extract_first_two_part_version(input_string: str) -> None:
-    cli_tool_audit.version_parsing.extract_first_two_part_version(
-        input_string=input_string
-    )
+    cli_tool_audit.version_parsing.extract_first_two_part_version(input_string=input_string)
 
 
 @given(input_string=st.text())